+++ cmd/gonotes/main.go
@@ -4,17 +4,22 @@ package main
 import (
 	"flag"
 	"fmt"
+	"os"
 
 	"github.com/mnbi/gonotes"
+	"github.com/mnbi/gonotes/internal/conf"
 )
 
 var (
 	versionFlag = flag.Bool("v", false, "show version")
 	usageFlag   = flag.Bool("h", false, "show usage")
+	configFile  string
 )
 
 func main() {
 	flag.Usage = gonotes.Usage
+	flag.StringVar(&configFile, "c", "", "specify a configuration file")
+
 	flag.Parse()
 
 	if *versionFlag {
@@ -32,4 +37,15 @@ func main() {
 		fmt.Printf(" %s", arg)
 	}
 	fmt.Printf("\n")
+
+	fmt.Printf("---- configuration settings ----\n")
+
+	if err := conf.Init(configFile); err != nil {
+		fmt.Fprint(os.Stderr, err)
+		os.Exit(1)
+	}
+
+	fmt.Printf("repo type: %s\n", conf.RepoType())
+	fmt.Printf("repo name: %s\n", conf.RepoName())
+	fmt.Printf("repo base: %s\n", conf.RepoBase())
 }