+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
-- Add `env`, which is a internal sub-package.
+- Add `conf`, which is a internal sub-package (#4).
+- Add `env`, which is a internal sub-package (#4).
 - Move `timestamp` into the `internal` directory.
 - Add `timestamp` sub-package (#2).
 - Add the build status badge to `README.md`